requirements.txt
numpy>=1.26.4
pandas>=2.2.0
matplotlib>=3.9.0
seaborn>=0.13.0
scikit-learn>=1.5.0

# TensorFlow with Python 3.12 support
tensorflow>=2.16.0

# Jupyter + kernel
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.29.0

# Advanced optimization and hyperparameter tuning
optuna>=3.4.0
hyperopt>=0.2.7

# Model interpretability
shap>=0.42.0
lime>=0.2.0

# Statistical analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Visualization enhancements
plotly>=5.17.0
plotly-express>=0.4.1

# Progress bars and utilities
tqdm>=4.66.0

# Data validation and quality
pandas-profiling>=3.6.0

# Advanced metrics and evaluation
imbalanced-learn>=0.11.0

# Configuration management
pyyaml>=6.0.1
configparser>=5.3.0

# Computer Vision & Image Processing for OCR
opencv-python>=4.8.0
pillow>=10.0.0

# OCR Libraries
pytesseract>=0.3.10
easyocr>=1.7.0

# Text Processing & Similarity
python-levenshtein>=0.20.9
nltk>=3.8.1

# Additional utilities for model evaluation
jinja2>=3.1.0

# Web API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
